+++ NEWS.md
@@ -1,4 +1,4 @@
-# pdp (development version)
+# pdp 0.9.1
 
 ## Deprecations